demo/agent_orchestrator/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
python-dotenv>=1.0.0
google-genai==1.10.0
httpx>=0.24.0
